requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
streamlit>=1.30.0
fastapi>=0.115.0
uvicorn>=0.30.0
google-generativeai>=0.6.0
